requirements.txt
scrapy
pyenchant
numpy
requests
Flask